DS4Windows no longer crashes if no source games are installed (ie shinra's crash) Auto profile has more ways to add programs, add origin games if installed, and start menu programs You can also add by shortcuts (.lnk) Auto Profiles now work by the foreground program DS4Windows no longer checks for driver at start, will help speed up DS4's connection when starting up with windows. Welcome to DS4Windows Window is now an option in the Settings Tab, will automatically show up if mine or DSDCS was never started before. Touchpad scrolling reworked via electrobrains newest build Idle Timeout has a checkbox, also small UI fixes

+++ DS4Control/MouseWheel.cs
@@ -33,8 +33,8 @@ namespace DS4Control
             Touch T1 = arg.touches[1];
 
             //mouse wheel 120 == 1 wheel click according to Windows API
-            int lastMidX = (lastT0.hwX + lastT1.hwX) / 2, lastMidY = (lastT0.hwY + lastT1.hwY) / 2,
-                currentMidX = (T0.hwX + T1.hwX) / 2, currentMidY = (T0.hwY + T1.hwY) / 2;
+            double lastMidX = (lastT0.hwX + lastT1.hwX) / 2d, lastMidY = (lastT0.hwY + lastT1.hwY) / 2d,
+               currentMidX = (T0.hwX + T1.hwX) / 2d, currentMidY = (T0.hwY + T1.hwY) / 2d;
             double coefficient = Global.getScrollSensitivity(deviceNumber);
             // Adjust for touch distance: "standard" distance is 960 pixels, i.e. half the width.  Scroll farther if fingers are farther apart, and vice versa, in linear proportion.
             double touchXDistance = T1.hwX - T0.hwX, touchYDistance = T1.hwY - T0.hwY, touchDistance = Math.Sqrt(touchXDistance * touchXDistance + touchYDistance * touchYDistance);
